backend/requirements.txt
# LangGraph + LangChain
langgraph>=0.2.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
aiosqlite>=0.20.0,<1.0.0

# LLM multi-provider
litellm>=1.50.0,<2.0.0

# Web framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sse-starlette>=2.0.0,<4.0.0
python-multipart>=0.0.9,<1.0.0

# Data validation & config
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# HTTP client
httpx>=0.27.0,<1.0.0

# Retry logic (used by jina_reader and google_search)
tenacity>=8.0.0,<10.0.0

# License module dependencies
cryptography>=43.0.0,<46.0.0
python-jose[cryptography]>=3.3.0,<4.0.0

# Observability
langfuse>=2.50.0,<3.0.0

# Search backends
tavily-python>=0.5.0,<1.0.0

# Document parsing
pymupdf4llm>=0.0.10,<1.0.0
openpyxl>=3.1.0,<4.0.0
pandas>=2.0.0,<3.0.0
python-docx>=1.1.0,<2.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.24.0,<1.0.0
pytest-cov>=5.0.0,<7.0.0
